+++ backend_api/app/applications/auth/auth_handler.py
@@ -1,5 +1,22 @@
+from applications.users.crud import get_user_by_email
+from settings import settings
+from fastapi import APIRouter, Depends
+from fastapi.security import OAuth2PasswordRequestForm
 
+from sqlalchemy.ext.asyncio import AsyncSession
+
+
+from database.session_dependancies import get_async_session
 
 class AuthHandler:
     def __init__(self):
-        pass
+        self.secret = settings.JWT_SECRET
+        self.algorithm = settings.JWT_ALGORITHM
+    async def get_login_token_pairs(self, data: OAuth2PasswordRequestForm, session: AsyncSession):
+        user_email = data.username
+        user_password = data.password
+        user = await get_user_by_email()
+        print(user, 888888)
+
+
+auth_handler = AuthHandler()
