+++ backend_api/app/applications/auth/router.py
@@ -3,7 +3,7 @@ from fastapi.security import OAuth2PasswordRequestForm
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-
+from applications.auth.auth_handler import auth_handler
 from database.session_dependancies import get_async_session
 
 
@@ -15,5 +15,5 @@ async def user_login(
     data: OAuth2PasswordRequestForm = Depends(),
     session: AsyncSession = Depends(get_async_session),
 ):
-
-    return
+    token_pair = await auth_handler.get_login_token_pairs(data, session)
+    return token_pair
